Types

Index

LimeSurveyBuilder.AbstractQuestion

LimeSurveyBuilder.AbstractSurveyComponent

LimeSurveyBuilder.LanguageSetting

LimeSurveyBuilder.LanguageSettings

LimeSurveyBuilder.Question

LimeSurveyBuilder.QuestionGroup

LimeSurveyBuilder.ResponseOption

LimeSurveyBuilder.ResponseScale

LimeSurveyBuilder.SubQuestion

LimeSurveyBuilder.Survey

Type specification

LimeSurveyBuilder.AbstractQuestion — Type

AbstractQuestion <: AbstractSurveyComponent

An abstract type representing a question within a LimeSurvey.

LimeSurveyBuilder.AbstractSurveyComponent — Type

AbstractSurveyComponent

An abstract type that represents a component within a LimeSurvey survey.

LimeSurveyBuilder.LanguageSetting — Type

LanguageSetting

A type representing the settings of a survey component for a single language.

Fields

language::String: The definition of the locale

title::title: The title of the survey component in the language

help::Union{Nothing, String}: The help text of the survey component in the language

description::Union{Nothing, String}: The description of the survey component in the language

default::Union{Nothing, String}: The default value of the survey component in the language

LimeSurveyBuilder.LanguageSettings — Type

LanguageSettings

A type representing a collection of single LanguageSetting.

Fields

settings::Vector{LanguageSetting}: A collection of language settings.

same_default::Bool: An indicator whether or not the same default value is used for all languages.

LimeSurveyBuilder.Question — Type

Question <: AbstractQuestion

A survey component that stores information about LimeSurvey questions.

Fields

id::String: An alphanumeric question id. Must start with a letter.

type::String: The LimeSurvey question type

mandatory::Bool: Determines if the question is mandatory

other::Bool: Determines if the questions as other category

relevance::String: A LimeSurvey Expression Script

attributes::Dict: Additional question attributes

language_settings::LanguageSettings: Language Settings for the question

subquestions::Vector{Subquestion}: A vector of subquestions

options::Vector{ResponseScale}: A vector of response scales.

LimeSurveyBuilder.QuestionGroup — Type

QuestionGroup <: AbstractSurveyComponent

A type representing a question group within a LimeSurvey

Fields

id::Int: An integer-valued ID for the question group

language_settings::LanguageSettings: The language settings for the question group

children::Vector{Question}: A vector of questions as child elements of the question group

LimeSurveyBuilder.ResponseOption — Type

ResponseOption <: AbstractSurveyComponent

A type representing a single response option for a LimeSurvey question.

Fields

id::String: A LimeSurvey ID for the response option

language_settings::LanguageSettings: The language settings for the response option

LimeSurveyBuilder.ResponseScale — Type

ResponseScale <: AbstractSurveyComponent

A type representing a collection of ResponseOption for a LimeSurvey question.

Fields

options::Vector{ResponseOption}: A vector of response options

language_settings::LanguageSettings: The language settings for the response scale.

LimeSurveyBuilder.SubQuestion — Type

SubQuestion <: AbstractQuestion

A type representing subquestions of a LimeSurvey question.

Fields

id::String: An alphanumeric question id. Must start with a letter.

question::String: The subquestion title.

type::String: The LimeSurvey qusetion type

relevance::String: A LimeSurvey Expression Script

LimeSurveyBuilder.Survey — Type

Survey <: AbstractSurveyComponent

A type to represent a LimeSurvey.

Fields

id::Int: A valid LimeSurvey survey id

language_settings::LanguageSettings: The surveys language settings

children::Vector{QuestionGroup}: A vector of question groups